% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance.R
\name{print.IndepBalance}
\alias{print.IndepBalance}
\title{Print an IndepBalance summary}
\usage{
\method{print}{IndepBalance}(x, ...)
}
\arguments{
\item{x}{An \code{IndepBalance} object from \code{check_balance()}.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print an IndepBalance summary
}
