% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.IndepAssoc}
\alias{summary.IndepAssoc}
\title{Summarize an IndepAssoc Result}
\usage{
\method{summary}{IndepAssoc}(object, ...)
}
\arguments{
\item{object}{An \code{IndepAssoc} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
The input \code{object}, invisibly.
}
\description{
Summarize an IndepAssoc Result
}
