% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests_stat.R
\name{mcnemar_test}
\alias{mcnemar_test}
\title{McNemar Test on Matched Data}
\usage{
mcnemar_test(matched_data, outcome, exposure)
}
\arguments{
\item{matched_data}{Matched data frame with \code{match_num} column.}

\item{outcome}{Character; binary outcome variable name.}

\item{exposure}{Character; exposure variable name.}
}
\value{
A data frame with McNemar test results.
}
\description{
McNemar Test on Matched Data
}
