% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome.R
\name{model_summ}
\alias{model_summ}
\title{Summarize a glm/lm/clogit model for the treatment effect}
\usage{
model_summ(model, treatment_feature, type = c("binary", "continuous"))
}
\description{
Summarize a glm/lm/clogit model for the treatment effect
}
\keyword{internal}
