% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome.R
\name{model_summ}
\alias{model_summ}
\title{Summarize a glm/lm/clogit model for the treatment effect}
\usage{
model_summ(model, treatment_feature, type = c("binary", "continuous"))
}
\description{
Summarize a glm/lm/clogit model for the treatment effect
}
\examples{
fit <- glm(outcome_binary ~ exposure + age,
           data = example_cohort, family = "binomial")
IndepAssoc:::model_summ(fit, "exposure", type = "binary")
}
\keyword{internal}
