% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests_stat.R
\name{paired_wilcoxon_test}
\alias{paired_wilcoxon_test}
\title{Paired Wilcoxon Test on Matched Data}
\usage{
paired_wilcoxon_test(matched_data, outcome, exposure)
}
\arguments{
\item{matched_data}{Matched data frame with \code{match_num} column.}

\item{outcome}{Character; continuous outcome variable name.}

\item{exposure}{Character; exposure variable name.}
}
\value{
A data frame with Wilcoxon test results.
}
\description{
Paired Wilcoxon Test on Matched Data
}
\examples{
data(example_cohort)
ps <- build_ps_model(example_cohort, "exposure",
                     c("age", "diabetes", "hypertension", "bmi"))
matched <- match_cohort(ps)
paired_wilcoxon_test(matched$data, "outcome_continuous", "exposure")

}
