% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.IndepAssoc}
\alias{print.IndepAssoc}
\title{Print an IndepAssoc Result}
\usage{
\method{print}{IndepAssoc}(x, ...)
}
\arguments{
\item{x}{An \code{IndepAssoc} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
The input \code{x}, invisibly.
}
\description{
Print an IndepAssoc Result
}
