% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{print.IndepAssoc}
\alias{print.IndepAssoc}
\title{Print an IndepAssoc pipeline result}
\usage{
\method{print}{IndepAssoc}(x, ...)
}
\arguments{
\item{x}{An \code{IndepAssoc} object from \code{run_pipeline()}.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print an IndepAssoc pipeline result
}
