% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{summary.IndepAssoc}
\alias{summary.IndepAssoc}
\title{Summary of an IndepAssoc pipeline result}
\usage{
\method{summary}{IndepAssoc}(object, ...)
}
\arguments{
\item{object}{An \code{IndepAssoc} object from \code{run_pipeline()}.}

\item{...}{Additional arguments passed to \code{print.IndepAssoc()}.}
}
\description{
Summary of an IndepAssoc pipeline result
}
